fix : Result의 역할을 LottoCalculator에 부여하다
